Update presstheattack.txt
I just realized the $ made the RegEx not perform as expected. The problem is that I tested it for simple domains such as `site.ninja` which it would work for but because of the $ it wont work if there is a / after the domain. Meaning, it could block `asuvadshj.club` but not `asuvadshj.club/asdasd.js`